Simpan hasil

+++ simpan_jawaban.php
@@ -1,25 +1,48 @@
 <?php
 include "koneksi.php";
-$id_siswa = $_POST['id_siswa'];
-$id_soal = $_POST['id_soal'];
-$hasil = $_POST['hasil'];
-$nama_file_code_siswa = "codesiswa_" . $id_siswa . "_" . $id_soal . ".php";
-$hapus_code_siswa = unlink("code/" . $nama_file_code_siswa);
-if($hapus_code_siswa){
-    $select_testcase = mysqli_query($koneksi, "SELECT * FROM testcase WHERE id_soal = '$id_soal'");
-    while($data_testcase = mysqli_fetch_assoc($select_testcase)){
-        $nama_file_testcase = "testcode_" . $id_siswa . "_" . $id_soal . "_" . $data_testcase["id_testcase"] . ".php";
-        $hapus_testcode = unlink("code/" . $nama_file_testcase);
-        if($hapus_testcode){
-            echo "testcode " . $nama_file_testcase . " berhasil dihapus<br>";
-        }
-        else {
-            echo "testcode gagal dihapus<br>";
-        }
-    }
 
-echo htmlspecialchars($hasil);
-}
-else {
-    echo "Gagal menghapus file";
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    // Mengambil dan menyaring input dari POST
+    $id_siswa = filter_input(INPUT_POST, 'id_siswa', FILTER_SANITIZE_NUMBER_INT);
+    $id_soal = filter_input(INPUT_POST, 'id_soal', FILTER_SANITIZE_NUMBER_INT);
+    $id_ujian = filter_input(INPUT_POST, 'id_ujian', FILTER_SANITIZE_NUMBER_INT);
+    $hasil = htmlspecialchars($_POST['hasil'], ENT_QUOTES, 'UTF-8');
+    $test_result = filter_input(INPUT_POST, 'test_result', FILTER_SANITIZE_STRING);
+
+    // Nama file code siswa
+    $nama_file_code_siswa = "codesiswa_" . $id_siswa . "_" . $id_soal . ".php";
+    $hapus_code_siswa = unlink("code/" . $nama_file_code_siswa);
+
+    if ($hapus_code_siswa) {
+        // Menyiapkan statement untuk mengambil data dari testcase
+        $stmt_testcase = $koneksi->prepare("SELECT id_testcase FROM testcase WHERE id_soal = ?");
+        $stmt_testcase->bind_param("i", $id_soal);
+        $stmt_testcase->execute();
+        $result_testcase = $stmt_testcase->get_result();
+
+        while ($data_testcase = $result_testcase->fetch_assoc()) {
+            $nama_file_testcase = "testcode_" . $id_siswa . "_" . $id_soal . "_" . $data_testcase["id_testcase"] . ".php";
+            $hapus_testcode = unlink("code/" . $nama_file_testcase);
+            if (!$hapus_testcode) {
+                echo "Gagal menghapus file testcase: " . $nama_file_testcase;
+                exit();
+            }
+        }
+
+        // Menyiapkan statement untuk menyimpan jawaban
+        $stmt_jawaban = $koneksi->prepare("INSERT INTO hasil (id_soal, id_siswa, jawaban, test_result) VALUES (?, ?, ?, ?)");
+        $stmt_jawaban->bind_param("iiss", $id_soal, $id_siswa, $hasil, $test_result);
+
+        if ($stmt_jawaban->execute()) {
+            // Menentukan lokasi home.php
+            $home_url = 'home.php?page=soal&id_ujian=' . $id_ujian;
+
+            // Script JavaScript untuk mengubah lokasi parent frame
+            echo "<script>window.parent.location.href = '$home_url';</script>";
+        } else {
+            echo "Gagal menyimpan jawaban";
+        }
+    } else {
+        echo "Gagal menghapus file";
+    }
 }
